requirements.txt
Eel==0.16.0
python-dotenv==1.0.1
